// src/payload-types.ts
export interface User {
  id: number
  name: string
  /**
   * URL friendly identifier
   */
  slug: string
  avatar?: (number | null) | Media
  job_title?: string | null
  bio?: string | null
  updatedAt: string
  createdAt: string
  enableAPIKey?: boolean | null
  apiKey?: string | null
  apiKeyIndex?: string | null
  email: string
  resetPasswordToken?: string | null
  resetPasswordExpiration?: string | null
  salt?: string | null
  hash?: string | null
  loginAttempts?: number | null
  lockUntil?: string | null
  sessions?:
    | {
        id: string
        createdAt?: string | null
        expiresAt: string
      }[]
    | null
  password?: string | null
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number
  alt: string
  folder?: ('posts' | 'avatars' | 'others') | null
  updatedAt: string
  createdAt: string
  url?: string | null
  thumbnailURL?: string | null
  filename?: string | null
  mimeType?: string | null
  filesize?: number | null
  width?: number | null
  height?: number | null
  focalX?: number | null
  focalY?: number | null
}
// ...
export interface Tag {
  id: number
  name: string
  slug: string
  updatedAt: string
  createdAt: string
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "subscribers".
 */
export interface Subscriber {
  id: number
  email: string
  source?: string | null
  updatedAt: string
  createdAt: string
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "topics".
 */
export interface Topic {
  id: number
  name: string
  display_name?: string | null
  slug: string
  active?: boolean | null
  /**
   * Must be reviewed before auto-publishing
   */
  requires_review?: boolean | null
  ai_status:
    | 'PENDING'
    | 'QUEUED'
    | 'METADATA_PROCESSING'
    | 'METADATA_COMPLETED'
    | 'METADATA_ERROR'
    | 'CONTENT_PROCESSING'
    | 'CONTENT_COMPLETED'
    | 'CONTENT_ERROR'
    | 'COMPLETED'
  /**
   * Click Share
   */
  click_share_sum?: number | null
  /**
   * Weighted Score
   */
  weighted_score_sum?: number | null
  alias?: string | null
  meta_title?: string | null
  meta_keywords?: string | null
  meta_description?: string | null
  excerpt?: string | null
  featured_media?: (number | null) | Media
  redirect?: boolean | null
  /**
   * Redirect URL, start with /
   */
  redirect_to?: string | null
  introductory?: {
    root: {
      type: string
      children: {
        type: any
        version: number
        [k: string]: unknown
      }[]
      direction: ('ltr' | 'rtl') | null
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | ''
      indent: number
      version: number
    }
    [k: string]: unknown
  } | null
  content?: {
    root: {
      type: string
      children: {
        type: any
        version: number
        [k: string]: unknown
      }[]
      direction: ('ltr' | 'rtl') | null
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | ''
      indent: number
      version: number
    }
    [k: string]: unknown
  } | null
  authors?: (number | User)[] | null
  tags?: (number | Tag)[] | null
  nodes_count?: number | null
  asin_count?: number | null
  /**
   * Conversion Share
   */
  conversion_share_sum?: number | null
  start_date?: string | null
  end_date?: string | null
  updatedAt: string
  createdAt: string
}
// ...
export interface AbaSearchTerm {
  id: number
  keywords: string
  asin: string
  /**
   * Node IDs
   */
  node_ids?:
    | {
        [k: string]: unknown
      }
    | unknown[]
    | string
    | number
    | boolean
    | null
  /**
   * Occurrences
   */
  occurrences?: number | null
  /**
   * Search Rank
   */
  search_rank?: number | null
  /**
   * Conversion Share
   */
  conversion_share?: number | null
  /**
   * Click Share
   */
  click_share?: number | null
  /**
   * Weighted Score
   */
  weighted_score?: number | null
  start_date?: string | null
  end_date?: string | null
  updatedAt: string
  createdAt: string
}